refactor: move contract state fns into mpt
