+++ crates/merkle-tree/Cargo.toml
@@ -14,6 +14,7 @@ pathfinder-storage = { path = "../storage" }
 rand = "0.8"
 rusqlite = { version = "0.27.0", features = ["bundled"] }
 stark_hash = { path = "../stark_hash" }
+starknet-gateway-types = { path = "../gateway-types" }
 
 [dev-dependencies]
 hex-literal = "0.3"
