+++ crates/merkle-tree/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod contract_state;
 pub mod merkle_node;
 pub mod merkle_tree;
 pub mod state_tree;
